backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
python-dotenv==1.0.0
firebase-admin==6.2.0
google-generativeai==0.3.2
transformers==4.35.2
torch>=2.2.0
huggingface-hub>=0.19.4
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
werkzeug==2.3.7
gunicorn==21.2.0
marshmallow==3.20.1
flasgger==0.9.7.1
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3
